Added 'position' parameter to `DrawPlane()`
This modification adds a new parameter called 'position' to the DrawPlane() function.
This parameter allows specifying the position where the plane should be drawn in the world, in addition to the origin, for greater ease of use.
I also increased the number of plane renders in the example for performance testing purposes.